[clang][dataflow] Move free functions out of DAContext.
